+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Added `prefix` to `Nullifier` (#1153).
 - [BREAKING] Implemented a `RemoteBatchProver`. `miden-proving-service` workers can prove batches (#1142).
 - [BREAKING] Implement `LocalBlockProver` and rename `Block` to `ProvenBlock` (#1152).
+- [BREAKING] Added native types to `AccountComponentTemplate` (#1124).
 
 ## 0.7.2 (2025-01-28) - `miden-objects` crate only